[Add]店舗一覧にレビューの評価合計と星に小数点追加

+++ db/migrate/20230424044338_create_reviews.rb
@@ -3,8 +3,8 @@ class CreateReviews < ActiveRecord::Migration[6.1]
     create_table :reviews do |t|
       t.references :customer, null: false, foreign_key: true
       t.references :space, null: false, foreign_key: true
-      t.integer     :rating, null: false, default: 0
-      t.text        :comment, null: false
+      t.float :rating, null: false, default: 0
+      t.text :comment, null: false
       
       t.index [:customer_id, :space_id], unique: true
       t.timestamps
